American National Election Studies (ANES) (2020) data

Based on the 2020 ANES data with derived variables and subset to people who completed both pre and post-election interviews

Usage

anes_2020

Format

A data frame with 7453 rows and 65 variables:

V200001

double 2020 Case ID

CaseID

double 2020 Case ID

V200002

double Mode of interview: pre-election interview

InterviewMode

factor Mode of interview: pre-election interview

V200010b

double Full sample post-election weight

Weight

double Full sample post-election weight

V200010c

double Full sample variance unit

VarUnit

factor Full sample variance unit

V200010d

double Full sample variance stratum

Stratum

factor Full sample variance stratum

V201006

double PRE: How interested in following campaigns

CampaignInterest

factor PRE: How interested in following campaigns

V201023

factor PRE: Confirmation voted (early) in November 3 Election (2020)

EarlyVote2020

factor PRE: Confirmation voted (early) in November 3 Election (2020)

V201024

double PRE: In what manner did R vote

V201025x

double PRE: SUMMARY: Registration and early vote status

V201028

double PRE: DID R VOTE FOR PRESIDENT

V201029

double PRE: For whom did R vote for President

V201101

double PRE: Did R vote for President in 2016 (revised)

V201102

double PRE: Did R vote for President in 2016

VotedPres2016

factor PRE: Did R vote for President in 2016

V201103

double PRE: Recall of last (2016) Presidential vote choice

VotedPres2016_selection

factor PRE: Recall of last (2016) Presidential vote choice

V201228

double PRE: Party ID: Does R think of self as Democrat, Republican, or Independent

V201229

double PRE: Party Identification strong - Democrat Republican

V201230

double PRE: No Party Identification - closer to Democratic Party or Republican Party

V201231x

double PRE: SUMMARY: Party ID

PartyID

factor PRE: SUMMARY: Party ID

V201233

double PRE: How often trust government in Washington to do what is right (revised)

TrustGovernment

factor PRE: How often trust government in Washington to do what is right (revised)

V201237

double PRE: How often can people be trusted

TrustPeople

factor PRE: How often can people be trusted

V201507x

double PRE: SUMMARY: Respondent age

Age

double PRE: SUMMARY: Respondent age

AgeGroup

factor PRE: SUMMARY: Respondent age (categorized)

V201510

double PRE: Highest level of Education

Education

factor PRE: Highest level of Education

V201546

double PRE: R: Are you Spanish, Hispanic, or Latino

V201547a

double RESTRICTED: PRE: Race of R: White (mention)

V201547b

double RESTRICTED: PRE: Race of R: Black or African-American (mention)

V201547c

double RESTRICTED: PRE: Race of R: Asian (mention)

V201547d

double RESTRICTED: PRE: Race of R: Native Hawaiian or Pacific Islander (mention)

V201547e

double RESTRICTED: PRE: Race of R: Native American or Alaska Native (mention)

V201547z

double RESTRICTED: PRE: Race of R: other specify

V201549x

double PRE: SUMMARY: R self-identified race/ethnicity

RaceEth

factor PRE: SUMMARY: R self-identified race/ethnicity

V201600

double PRE: What is your (R) sex? (revised)

Gender

factor PRE: What is your (R) sex? (revised)

V201607

double RESTRICTED: PRE: Total income amount - revised

V201610

double RESTRICTED: PRE: Income amt missing - categories lt 20K

V201611

double RESTRICTED: PRE: Income amt missing - categories 20-40K

V201613

double RESTRICTED: PRE: Income amt missing - categories 40-70K

V201615

double RESTRICTED: PRE: Income amt missing - categories 70-100K

V201616

double RESTRICTED: PRE: Income amt missing - categories 100+K

V201617x

double PRE: SUMMARY: Total (family) income

Income

factor PRE: SUMMARY: Total (family) income (21 categories)

Income7

factor PRE: SUMMARY: Total (family) income (7 categories)

V202051

double POST: R registered to vote (post-election)

V202066

double POST: Did R vote in November 2020 election

V202072

double POST: Did R vote for President

VotedPres2020

factor POST: Did R vote for President

V202073

double POST: For whom did R vote for President

V202109x

double PRE-POST: SUMMARY: Voter turnout in 2020

V202110x

double PRE-POST: SUMMARY: 2020 Presidential vote

VotedPres2020_selection

factor PRE-POST: SUMMARY: 2020 Presidential vote

Source

https://electionstudies.org/data-center/2020-time-series-study/