Residential Energy Consumption Survey (RECS) (2015) data

Based on the 2015 RECS microdata with derived variables

Usage

recs_2015

Format

A data frame with 5686 rows and 141 variables:

DOEID

double Unique identifier for each respondent

REGIONC

double Census Region (numeric)

Region

factor Census Region (factor)

Division

factor Census Division

MSAStatus

factor Housing unit in Census Metropolitan Statistical Area or Micropolitan Statistical Area

Urbanicity

factor Census 2010 Urban Type

HousingUnitType

factor Type of housing unit

YearMade

factor Year range when housing unit was built

SpaceHeatingUsed

logical Space heating used

HeatingBehavior

factor Main heating equipment household behavior

WinterTempDay

double Winter temperature when someone is home during the day

WinterTempAway

double Winter temperature when no one is home during the day

WinterTempNight

double Winter temperature at night

ACUsed

logical Air conditioning equipment used

ACBehavior

factor Central air conditioner household behavior

SummerTempDay

double Summer temperature when someone is home during the day

SummerTempAway

double Summer temperature when no one is home during the day

SummerTempNight

double Summer temperature at night

TOTCSQFT

double Total cooled square footage

TOTHSQFT

double Total heated square footage

TOTSQFT_EN

double Total square footage (used for publication)

TOTUCSQFT

double Total uncooled square footage

TOTUSQFT

double Total unheated square footage

NWEIGHT

double Final sample weight

BRRWT1 - BRRWT96

double Replicate weight 1 through 96

CDD30YR

double Cooling degree days, 30-year average 1981-2010, base temperature 65F

CDD65

double Cooling degree days in 2015, base temperature 65F

CDD80

double Cooling degree days in 2015, base temperature 80F (used for garage cooling load estimation only)

ClimateRegion_BA

factor Building America Climate Zone

ClimateRegion_IECC

factor IECC Climate Code

HDD30YR

double Heating degree days, 30-year average 1981-2010, base temperature 65F

HDD65

double Heating degree days in 2015, base temperature 65F

HDD50

double Heating degree days in 2015, base temperature 50F (used for garage heating load estimation only)

GNDHDD65

double Heating degree days of ground temperature in 2015, base temperature 65F

BTUEL

double Total site electricity usage, in thousand Btu, 2015

DOLLAREL

double Total electricity cost, in dollars, 2015

BTUNG

double Total natural gas usage, in thousand Btu, 2015

DOLLARNG

double Total natural gas cost, in dollars, 2015

BTULP

double Total propane usage, in thousand Btu, 2015

DOLLARLP

double Total cost of propane, in dollars, 2015

BTUFO

double Total fuel oil/kerosene usage, in thousand Btu, 2015

DOLLARFO

double Total cost of fuel oil/kerosene, in dollars, 2015

TOTALBTU

double Total usage, in thousand Btu, 2015

TOTALDOL

double Total cost, in dollars, 2015

BTUWOOD

double Total cordwood usage, in thousand Btu, 2015 (Wood consumption is not included in TOTALBTU or TOTALDOL)

BTUPELLET

double Total wood pellet usage, in thousand Btu, 2015 (Wood consumption is not included in TOTALBTU or TOTALDOL)

Source

https://www.eia.gov/consumption/residential/data/2015/index.php?view=microdata