Residential Energy Consumption Survey (RECS) (2020) data

Based on the 2020 RECS microdata with derived variables

Usage

recs_2020

Format

A data frame with 18496 rows and 118 variables:

DOEID

double Unique identifier for each respondent

ClimateRegion_BA

factor Building America Climate Zone

Urbanicity

factor 2010 Census Urban Type Code

Region

factor Census Region

REGIONC

character Census Region

Division

factor Census Division, Mountain Division is divided into North and South for RECS purposes

STATE_FIPS

character State Federal Information Processing System Code

state_postal

factor State Postal Code

state_name

factor State Name

HDD65

double Heating degree days in 2020, base temperature 65F; Derived from the weighted temperatures of nearby weather stations

CDD65

double Cooling degree days in 2020, base temperature 65F; Derived from the weighted temperatures of nearby weather stations

HDD30YR

double Heating degree days, 30-year average 1981-2010, base temperature 65F; Taken from nearest weather station, inoculated with random errors

CDD30YR

double Cooling degree days, 30-year average 1981-2010, base temperature 65F; Taken from nearest weather station, inoculated with random errors

HousingUnitType

factor Type of housing unit

YearMade

factor Range when housing unit was built

TOTSQFT_EN

double Total energy-consuming area (square footage) of the housing unit. Includes all main living areas; all basements; heated, cooled, or finished attics; and heating or cooled garages. For single-family housing units this is derived using the respondent-reported square footage (SQFTEST) and adjusted using the "include" variables (e.g., SQFTINCB), where applicable. For apartments and mobile homes this is the respondent-reported square footage. A derived variable rounded to the nearest 10

TOTHSQFT

double Square footage of the housing unit that is heated by space heating equipment. A derived variable rounded to the nearest 10

TOTCSQFT

double Square footage of the housing unit that is cooled by air-conditioning equipment or evaporative cooler, a derived variable rounded to the nearest 10

SpaceHeatingUsed

logical Space heating equipment used

ACUsed

logical Air conditioning equipment used

HeatingBehavior

factor Winter temperature control method

WinterTempDay

double Winter thermostat setting or temperature in home when someone is home during the day

WinterTempAway

double Winter thermostat setting or temperature in home when no one is home during the day

WinterTempNight

double Winter thermostat setting or temperature in home at night

ACBehavior

factor Summer temperature control method

SummerTempDay

double Summer thermostat setting or temperature in home when someone is home during the day

SummerTempAway

double Summer thermostat setting or temperature in home when no one is home during the day

SummerTempNight

double Summer thermostat setting or temperature in home at night

NWEIGHT

double Final Analysis Weight

NWEIGHT1 - NWEIGHT60

double Final Analysis Weight for replicate 1-60

BTUEL

double Total electricity use, in thousand Btu, 2020, including self-generation of solar power

DOLLAREL

double Total electricity cost, in dollars, 2020

BTUNG

double Total natural gas use, in thousand Btu, 2020

DOLLARNG

double Total natural gas cost, in dollars, 2020

BTULP

double Total propane use, in thousand Btu, 2020

DOLLARLP

double Total propane cost, in dollars, 2020

BTUFO

double Total fuel oil/kerosene use, in thousand Btu, 2020

DOLLARFO

double Total fuel oil/kerosene cost, in dollars, 2020

BTUWOOD

double Total wood use, in thousand Btu, 2020

TOTALBTU

double Total usage including electricity, natural gas, propane, and fuel oil, in thousand Btu, 2020

TOTALDOL

double Total cost including electricity, natural gas, propane, and fuel oil, in dollars, 2020

Source

https://www.eia.gov/consumption/residential/data/2020/index.php?view=microdata